refactor(wrapped-keys-lit-actions): LIT-3959 - Extract `exportPrivateKey` from IIFE
